LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/json-module/module.tentative-expected.txt

CONSOLE MESSAGE: TypeError: 'application/json' is not a valid JavaScript MIME type.

FAIL JSON modules assert_unreached: [object Event] Reached unreachable code
